requirements.txt
Django==1.10
django-debug-toolbar==1.5
django-extensions==1.7.1
psycopg2==2.6.2
six==1.10.0
sqlparse==0.2.0
python-docx==0.8.6